modificacion barra Privada y logout
